# apps/web/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditDetailsResponse": {
        "properties": {
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "rule_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule Id"
          }
        },
        "required": [
          "count",
          "rule_id"
        ],
        "title": "AuditDetailsResponse",
        "type": "object"
      },
      "AuditEntityRefResponse": {
        "properties": {
          "entity_id": {
            "format": "uuid",
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "entity_type",
          "entity_id",
          "version"
        ],
        "title": "AuditEntityRefResponse",
        "type": "object"
      },
      "AuditEventResponse": {
        "properties": {
          "actor_type": {
            "title": "Actor Type",
            "type": "string"
          },
          "audit_event_id": {
            "format": "uuid",
            "title": "Audit Event Id",
            "type": "string"
          },
          "causation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Causation Id"
          },
          "correlation_id": {
            "format": "uuid",
            "title": "Correlation Id",
            "type": "string"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "entity_refs": {
            "items": {
              "$ref": "#/components/schemas/AuditEntityRefResponse"
            },
            "title": "Entity Refs",
            "type": "array"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "input_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Hash"
          },
          "occurred_at": {
            "title": "Occurred At",
            "type": "string"
          },
          "output_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Hash"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "safe_details": {
            "$ref": "#/components/schemas/AuditDetailsResponse"
          }
        },
        "required": [
          "audit_event_id",
          "event_type",
          "occurred_at",
          "actor_type",
          "decision",
          "reason_codes",
          "entity_refs",
          "safe_details",
          "correlation_id",
          "causation_id",
          "input_hash",
          "output_hash"
        ],
        "title": "AuditEventResponse",
        "type": "object"
      },
      "AuditPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "AuditPageResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ambient CHORUS",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cases/{case_id}/audit": {
      "get": {
        "description": "Page one case's safe audit trail, presenter-only, in occurrence order.",
        "operationId": "read_case_audit_v1_cases__case_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "X-Chorus-Demo-Actor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Chorus-Demo-Actor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Case Audit",
        "tags": [
          "audit"
        ]
      }
    },
...
  }
}